added more movement, see extd descrip
The last key to this hell of a project is linking the way we talk to the computer in terms of our z axis and direction dictionaries. I am confident that after that is solved, we will have maybe a small little bug here or there, but ultimately will have motion fixed.